Move Contribution model to own app
