+++ backend/app/settings.py
@@ -63,6 +63,7 @@ THIRD_PARTY_APPS = [
 LOCAL_APPS = [
     "core",
     "api",
+    "contributions",
     "stats",
 ]
