+++ frontend/src/components/FeedbackCard.vue
@@ -123,7 +123,7 @@ export default {
       this.contributionSubmitted = true;
       this.error = this.contributionResponse = null;
       this.loading = true;
-      fetch(`${process.env.VUE_APP_API_ENDPOINT}/contribute`, {
+      fetch(`${process.env.VUE_APP_API_ENDPOINT}/contributions`, {
         method: 'POST',
         headers: {
           Accept: 'application/json',
